+++ docs/src/global/app.ts
@@ -19,3 +19,4 @@ import '@deckdeckgo/highlight-code';
 import '@deckdeckgo/charts';
 import '@deckdeckgo/qrcode';
 import '@deckdeckgo/inline-editor';
+import '@deckdeckgo/color';