+++ studio/src/global/app-dev.ts
@@ -12,6 +12,7 @@ import '@deckdeckgo/remote';
 import '@deckdeckgo/qrcode';
 import '@deckdeckgo/highlight-code';
 import '@deckdeckgo/lazy-img';
+import '@deckdeckgo/color';
 
 import '@deckdeckgo/slide-title';
 import '@deckdeckgo/slide-content';